GDGT-2684: Move old SidebarResizableBox to ModelSibar
