+++ enterprise/frontend/src/metabase-enterprise/replacement/pages/MigrateModelsPage/ModelSidebar/ModelSidebar.tsx
@@ -3,7 +3,6 @@ import { memo } from "react";
 import { skipToken, useGetCardQuery, useGetDatabaseQuery } from "metabase/api";
 import { DelayedLoadingAndErrorWrapper } from "metabase/common/components/LoadingAndErrorWrapper/DelayedLoadingAndErrorWrapper";
 import { Box, Center, Stack } from "metabase/ui";
-import { SidebarResizableBox } from "metabase-enterprise/monitor/dependency-diagnostics/components/DiagnosticsSidebar/SidebarResizableBox";
 import type { CardId } from "metabase-types/api";
 
 import { ActionSection } from "./ActionSection";
@@ -12,6 +11,7 @@ import { InfoSection } from "./InfoSection";
 import { LocationSection } from "./LocationSection";
 import S from "./ModelSidebar.module.css";
 import { SidebarHeader } from "./SidebarHeader";
+import { SidebarResizableBox } from "./SidebarResizableBox";
 
 type ModelSidebarProps = {
   cardId: CardId;
