+++ enterprise/frontend/src/metabase-enterprise/replacement/pages/MigrateModelsPage/ModelSidebar/SidebarResizableBox/SidebarResizableBox.tsx
